Store cost columns as double precision to match the mapping

+++ dokimos-server/src/main/resources/db/migration/V9__item_metrics.sql
@@ -3,12 +3,12 @@
 -- metrics stay valid; no backfill.
 ALTER TABLE item_results ADD COLUMN tokens_in INTEGER;
 ALTER TABLE item_results ADD COLUMN tokens_out INTEGER;
-ALTER TABLE item_results ADD COLUMN cost_usd NUMERIC(18, 8);
+ALTER TABLE item_results ADD COLUMN cost_usd DOUBLE PRECISION;
 ALTER TABLE item_results ADD COLUMN latency_ms BIGINT;
 
 -- Run-level aggregates materialized at run completion, mirroring the existing pass-rate fields.
 -- All nullable; absent means no item in the run carried that metric.
 ALTER TABLE experiment_runs ADD COLUMN total_tokens_in BIGINT;
 ALTER TABLE experiment_runs ADD COLUMN total_tokens_out BIGINT;
-ALTER TABLE experiment_runs ADD COLUMN total_cost_usd NUMERIC(18, 8);
+ALTER TABLE experiment_runs ADD COLUMN total_cost_usd DOUBLE PRECISION;
 ALTER TABLE experiment_runs ADD COLUMN avg_latency_ms DOUBLE PRECISION;
